creacion de la estructura MVVM
