+++ app/src/main/java/com/example/proyectoshopifyka/utils/FragmentComunicator.kt
@@ -1,0 +1,5 @@
+package com.example.proyectoshopifyka.utils
+
+interface FragmentComunicator {
+    fun showLoader(value: Boolean)
+}
