+++ app/src/main/java/com/example/proyectoshopifyka/view/FirstFragment.kt
@@ -1,3 +1,4 @@
+
 package com.example.proyectoshopifyka
 
 import android.os.Bundle
@@ -7,7 +8,6 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
 import com.example.proyectoshopifyka.databinding.FragmentFirstBinding
-
 /**
  * A simple [Fragment] subclass as the default destination in the navigation.
  */
