+++ app/src/main/java/com/example/proyectoshopifyka/view/MainActivity2.kt
@@ -9,6 +9,7 @@ import androidx.navigation.ui.navigateUp
 import androidx.navigation.ui.setupActionBarWithNavController
 import com.example.proyectoshopifyka.databinding.ActivityMain2Binding
 
+
 class MainActivity2 : AppCompatActivity() {
 
     private lateinit var appBarConfiguration: AppBarConfiguration
