+++ app/src/main/java/com/example/proyectoshopifyka/view/PersonalInformationVariant.kt
@@ -1,3 +1,4 @@
+
 package com.example.proyectoshopifyka
 
 import android.os.Bundle
@@ -6,16 +7,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
 
-/**
- * A simple [Fragment] subclass.
- * Use the [PersonalInformationVariant.newInstance] factory method to
- * create an instance of this fragment.
- */
 class PersonalInformationVariant : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
@@ -24,8 +16,7 @@ class PersonalInformationVariant : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
+
         }
     }
 
@@ -51,8 +42,7 @@ class PersonalInformationVariant : Fragment() {
         fun newInstance(param1: String, param2: String) =
             PersonalInformationVariant().apply {
                 arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
+
                 }
             }
     }
