+++ app/src/main/java/com/example/proyectoshopifyka/view/layout_register.kt
@@ -6,16 +6,6 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-/**
- * A simple [Fragment] subclass.
- * Use the [layout_register.newInstance] factory method to
- * create an instance of this fragment.
- */
 class layout_register : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
@@ -24,8 +14,7 @@ class layout_register : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
+
         }
     }
 
@@ -51,8 +40,7 @@ class layout_register : Fragment() {
         fun newInstance(param1: String, param2: String) =
             layout_register().apply {
                 arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
+
                 }
             }
     }
