+++ app/src/main/java/com/example/proyectoshopifyka/viewModel/SignInViewModel.kt
@@ -1,0 +1,4 @@
+package com.example.proyectoshopifyka.viewModel
+
+class SignInViewModel {
+}
